use mimetype to decide if typesetting is needed (#38)
Use mimetype to decide if typesetting is needed

When the path portion of the URI contains . - for example in the query part
like in `geminispace.info/search?test.com` - the page is not handled by `typesetgmi`.
Instead of matching the `path` part of the URI we check the mimetype from meta instead.

I'm currently unsure if the dot in a query part should be percent-encoded anyway (quick research
lead to ambiguous results). At least for now astro does not percent-encode it.